Initial bad account or duplicate account code implement
The initial bad account or duplicate account error checking implemented.  Also, updated the database structure so only the name field exists in the database instead of the first name and last name fields.  Updated the config database scripts and account functions to handle the later.

Updated the header to display the person's name.